+++ openid-data-classes/src/commonMain/kotlin/at/asitplus/wallet/lib/data/rfc/tokenStatusList/iso18013/RFU.kt
@@ -1,4 +1,4 @@
-package at.asitplus.wallet.lib.data.iso18013
+package at.asitplus.wallet.lib.data.rfc.tokenStatusList.iso18013
 
 import kotlinx.serialization.Serializable
